move ex enum to types
