+++ src/components/content/exercises/exercise.tsx
@@ -3,7 +3,6 @@ import dynamic from 'next/dynamic'
 import { useState, useEffect } from 'react'
 
 import { LicenseNotice } from '../license/license-notice'
-import { ExerciseInlineType } from './exercise-author-tools'
 import { ExerciseNumbering } from './exercise-numbering'
 import { InputExercise } from './input-exercise'
 import { ScMcExercise } from './sc-mc-exercise'
@@ -13,6 +12,7 @@ import { Lazy } from '@/components/content/lazy'
 import { isPrintMode, printModeSolutionVisible } from '@/components/print-mode'
 import { useInstanceData } from '@/contexts/instance-context'
 import { useLoggedInComponents } from '@/contexts/logged-in-components'
+import { ExerciseInlineType } from '@/data-types'
 import { FrontendExerciseNode, FrontendNodeType } from '@/frontend-node-types'
 import type { NodePath, RenderNestedFunction } from '@/schema/article-renderer'
