+++ src/components/user-tools/author-tools-hover-menu.tsx
@@ -2,10 +2,9 @@ import Tippy, { TippyProps } from '@tippyjs/react'
 
 import { AuthorTools, Tool } from './author-tools'
 import { MenuSubButtonLink } from './menu-sub-button-link'
-import { ExerciseInlineType } from '@/components/content/exercises/exercise-author-tools'
 import { useInstanceData } from '@/contexts/instance-context'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
-import { UuidType, UuidWithRevType } from '@/data-types'
+import { ExerciseInlineType, UuidType, UuidWithRevType } from '@/data-types'
 import { TaxonomyTermType } from '@/fetcher/graphql-types/operations'
 import { getTranslatedType } from '@/helper/get-translated-type'
