+++ src/components/user-tools/author-tools.tsx
@@ -3,13 +3,12 @@ import Tippy from '@tippyjs/react'
 import { useRouter } from 'next/router'
 import { Fragment } from 'react'
 
-import { ExerciseInlineType } from '../content/exercises/exercise-author-tools'
 import { AuthorToolsData, tippyDefaultProps } from './author-tools-hover-menu'
 import { MenuSubButtonLink } from './menu-sub-button-link'
 import { useCanDo } from '@/auth/use-can-do'
 import { useInstanceData } from '@/contexts/instance-context'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
-import { UuidRevType, UuidType } from '@/data-types'
+import { ExerciseInlineType, UuidRevType, UuidType } from '@/data-types'
 import { Instance, TaxonomyTermType } from '@/fetcher/graphql-types/operations'
 import { getTranslatedType } from '@/helper/get-translated-type'
 import { useSetUuidStateMutation } from '@/helper/mutations/use-set-uuid-state-mutation'
