+++ src/components/user-tools/user-tools.tsx
@@ -9,7 +9,6 @@ import { TaxonomyTerm, Uuid } from '@serlo/authorization'
 import clsx from 'clsx'
 import { useEffect, useState, cloneElement } from 'react'
 
-import { ExerciseInlineType } from '../content/exercises/exercise-author-tools'
 import { FaIcon } from '../fa-icon'
 import { LazyTippy } from '../navigation/lazy-tippy'
 import { AuthorToolsData } from './author-tools-hover-menu'
@@ -19,7 +18,7 @@ import { Link } from '@/components/content/link'
 import { useInstanceData } from '@/contexts/instance-context'
 import { useLoggedInComponents } from '@/contexts/logged-in-components'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
-import { UuidRevType, UuidType } from '@/data-types'
+import { ExerciseInlineType, UuidRevType, UuidType } from '@/data-types'
 import { Instance } from '@/fetcher/graphql-types/operations'
 import { getEditUrl } from '@/helper/urls/get-edit-url'
 import { getHistoryUrl } from '@/helper/urls/get-history-url'
